Clean up, copyrights, and corrections
Remove system out

+++ dev/com.ibm.ws.transport.http/src/com/ibm/ws/http/channel/internal/HttpChannelConfig.java
@@ -854,7 +854,7 @@ public class HttpChannelConfig {
 
     }
 
-    public long getWebSocketBufferSize() {
+    public int getWebSocketBufferSize() {
         return this.websocketBufferSize;
     }
